+++ src/assets/blocks/coming-soon/edit.js
@@ -35,19 +35,19 @@ export default function Edit ( { attributes, setAttributes } ) {
 				</div>
 
 				<InspectorControls>
-					<PanelBody title={__('Settings', 'lumiere-movies')}>
+					<PanelBody title={ __('Main settings', 'lumiere-movies')}>
 						<SelectControl
 							label={ __( 'Country', 'lumiere-movies' ) }
-							value={attributes.region}
-							onChange={region => setAttributes({ region })}
+							value={ attributes.region }
+							onChange={ region => setAttributes( { region } ) }
 							__nextHasNoMarginBottom
 							__next40pxDefaultSize
 							options={ imdbCountries }
 						/>
 						<SelectControl
 							label={ __( 'Type of search', 'lumiere-movies' ) }
-							value={attributes.type}
-							onChange={type => setAttributes({ type })}
+							value={ attributes.type }
+							onChange={ type => setAttributes( { type } ) }
 							__nextHasNoMarginBottom
 							__next40pxDefaultSize
 							options={ [
@@ -56,24 +56,45 @@ export default function Edit ( { attributes, setAttributes } ) {
 								{ label: 'TV episodes', value: 'TV_EPISODE' },
 							] }
 						/>
-						<TextControl
-							label={__('Starting date (in number of days)', 'lumiere-movies')}
-							type="number"
-							value={attributes.startDateOverride}
-							onChange={val => setAttributes({ startDateOverride: parseInt(val) || 0 })}
-							__nextHasNoMarginBottom={ true }
-							__next40pxDefaultSize={ true }
-						/>
-						<TextControl
-							label={__('Ending date (in number of days)', 'lumiere-movies')}
-							type="number"
-							value={attributes.endDateOverride}
-							onChange={val => setAttributes({ endDateOverride: parseInt(val) || 0 })}
-							__nextHasNoMarginBottom={ true }
-							__next40pxDefaultSize={ true }
-						/>
 					</PanelBody>
 				</InspectorControls>
+				<InspectorControls>
+					<PanelBody
+						title={ __( 'Date settings', 'lumiere-movies' ) }
+						initialOpen={ false }
+					>
+						<TextControl
+							label={ __( 'Starting date (in number of days)', 'lumiere-movies' ) }
+							type="number"
+							value={ attributes.startDateOverride }
+							onChange={val => setAttributes( { startDateOverride: parseInt(val) || 0 } ) }
+							__nextHasNoMarginBottom={ true }
+							__next40pxDefaultSize={ true }
+						/>
+						<TextControl
+							label={ __( 'Ending date (in number of days)', 'lumiere-movies' ) }
+							type="number"
+							value={ attributes.endDateOverride }
+							onChange={ val => setAttributes( { endDateOverride: parseInt(val) || 0 } ) }
+							__nextHasNoMarginBottom={ true }
+							__next40pxDefaultSize={ true }
+						/>
+						<SelectControl
+							label={ __( 'Date format', 'lumiere-movies' ) }
+							type="text"
+							value={ attributes.dateFormatOverride }
+							onChange={ dateFormatOverride => setAttributes( { dateFormatOverride } ) }
+							__nextHasNoMarginBottom
+							__next40pxDefaultSize
+							options={ [
+								{ label: __( 'Blog default format', 'lumiere-movies' ), value: 'WordPress format' },
+								{ label: __( 'Full format', 'lumiere-movies' ), value: 'l d F Y' },
+								{ label: __( 'Partial format', 'lumiere-movies' ), value: 'D d M Y' },
+								{ label: __( 'Short format', 'lumiere-movies' ), value: 'd/m/Y' }
+							] }
+						/>
+					</PanelBody>
+				 </InspectorControls>
 			</div>
 			<ServerSideRender
 				block={ jsonData.name }