+++ src/assets/blocks/widget/edit.js
@@ -1,5 +1,5 @@
 import { useBlockProps } from '@wordpress/block-editor';
-import { __ } from '@wordpress/i18n';
+import { __, sprintf } from '@wordpress/i18n';
 import { RawHTML } from '@wordpress/element';
 
 export default function Edit ( { attributes, setAttributes } ) {